requirements.txt
psutil
Flask
google-generativeai
python-dotenv
colorama
rich
python-dotenv
